add display

+++ tuyaos/tuyaos_adapter/include/spi/tkl_spi.h
@@ -45,7 +45,7 @@ OPERATE_RET tkl_spi_deinit(TUYA_SPI_NUM_E port);
  *
  * @return  OPRT_OK on success. Others on error, please refer to tuya_error_code.h
  */
-OPERATE_RET tkl_spi_send(TUYA_SPI_NUM_E port, void *data, uint16_t size);
+OPERATE_RET tkl_spi_send(TUYA_SPI_NUM_E port, void *data,  uint32_t size);
 
 /**
  * spi_recv
